requirements.txt
discord.py>=2.5.0
python-dotenv>=1.0.0
openai>=1.8.0
requests>=2.30.0 